Improve the fix for the theory data type cast issue. (Don't manipulate TestMethodArguments value any more)
