+++ src/xunit.execution/Sdk/Frameworks/Runners/TestInvoker.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Security;
@@ -143,9 +144,10 @@ namespace Xunit.Sdk
         /// if you need to do some other form of invocation of the actual test method.
         /// </summary>
         /// <param name="testClassInstance">The instance of the test class</param>
+        /// <param name="arguments">The arguments the testmethod should be called with</param>
         /// <returns>The return value from the test method invocation</returns>
-        protected virtual object CallTestMethod(object testClassInstance)
-            => TestMethod.Invoke(testClassInstance, TestMethodArguments);
+        protected virtual object CallTestMethod(object testClassInstance, object[] arguments)
+            => TestMethod.Invoke(testClassInstance, arguments);
 
         /// <summary>
         /// Creates the test class (if necessary), and invokes the test method.
@@ -223,13 +225,8 @@ namespace Xunit.Sdk
                             }
                             else
                             {
-                                for (int i = 0; i < parameterCount; i++)
-                                {
-                                    if (TestMethodArguments[i] != null && parameters[i].ParameterType != TestMethodArguments[i].GetType())
-                                        TestMethodArguments[i] = Cast(TestMethodArguments[i], parameters[i].ParameterType);
-                                }
-
-                                var result = CallTestMethod(testClassInstance);
+                                var result = CallTestMethod(testClassInstance, 
+                                    TestMethodArguments.Select((x, i) => Cast(x, parameters[i].ParameterType)).ToArray());
                                 var task = result as Task;
                                 if (task != null)
                                     await task;
@@ -257,6 +254,6 @@ namespace Xunit.Sdk
             => SynchronizationContext.SetSynchronizationContext(context);
 
         static object Cast(object value, Type targetType)
-            => Expression.Lambda<Func<object>>(Expression.Convert(Expression.Constant(value), targetType)).Compile()();
+            => value == null || value.GetType() == targetType ? value : Expression.Lambda<Func<object>>(Expression.Convert(Expression.Constant(value), targetType)).Compile()();
     }
 }
