+++ test/test.xunit.execution/test.xunit.execution.csproj
@@ -86,6 +86,7 @@
     <Compile Include="Acceptance\FixtureAcceptanceTests.cs" />
     <Compile Include="Acceptance\Xunit2AcceptanceTests.cs" />
     <Compile Include="Acceptance\Xunit2TheoryAcceptanceTests.cs" />
+    <Compile Include="TheoryDataTypeTests.cs" />
     <Compile Include="Common\AssemblyExtensionsTests.cs" />
     <Compile Include="Common\SerializationHelperTests.cs" />
     <Compile Include="Common\TraitHelperTests.cs" />
